+++ docs/html/search/all_0.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['aceroutine_20library_0',['AceRoutine Library',['../index.html',1,'']]]
+  ['ace_5froutine_5fdeprecated_0',['ACE_ROUTINE_DEPRECATED',['../Coroutine_8h.html#aa4ab589f43432a7c08a770e4c6f9b239',1,'Coroutine.h']]],
+  ['aceroutine_20library_1',['AceRoutine Library',['../index.html',1,'']]]
 ];