+++ docs/html/search/all_9.js
@@ -1,19 +1,7 @@
 var searchData=
 [
-  ['setdelaying_52',['setDelaying',['../classace__routine_1_1Coroutine.html#aa7db4b21985bbed9996427179d9a6dba',1,'ace_routine::Coroutine']]],
-  ['setdelaymicros_53',['setDelayMicros',['../classace__routine_1_1Coroutine.html#a5cf7c47f82ca67910955e4a9c2c7afe9',1,'ace_routine::Coroutine']]],
-  ['setdelaymillis_54',['setDelayMillis',['../classace__routine_1_1Coroutine.html#af89fa38dbebbd7b4ab31d5af821e3da6',1,'ace_routine::Coroutine']]],
-  ['setdelayseconds_55',['setDelaySeconds',['../classace__routine_1_1Coroutine.html#a63564eb4165b40c40ce6d44c63ef78b8',1,'ace_routine::Coroutine']]],
-  ['setending_56',['setEnding',['../classace__routine_1_1Coroutine.html#a5dd0810009383cdeae859fc227d933bb',1,'ace_routine::Coroutine']]],
-  ['setjump_57',['setJump',['../classace__routine_1_1Coroutine.html#a11b968a5aeaa08acb19a684108c13805',1,'ace_routine::Coroutine']]],
-  ['setrunning_58',['setRunning',['../classace__routine_1_1Coroutine.html#a32e726c256adc0dce64819c8cf9a38c2',1,'ace_routine::Coroutine']]],
-  ['setterminated_59',['setTerminated',['../classace__routine_1_1Coroutine.html#a9c3fa246606d04d205eb4b0c2e91406c',1,'ace_routine::Coroutine']]],
-  ['setup_60',['setup',['../classace__routine_1_1CoroutineScheduler.html#a7a1e65b27de84506ed1427c240314ade',1,'ace_routine::CoroutineScheduler']]],
-  ['setupcoroutine_61',['setupCoroutine',['../classace__routine_1_1Coroutine.html#a3417537c1ff3a7c737719d32192f3733',1,'ace_routine::Coroutine::setupCoroutine(const char *name)'],['../classace__routine_1_1Coroutine.html#a68dc92a6cce8435bc745a3582bbca210',1,'ace_routine::Coroutine::setupCoroutine(const __FlashStringHelper *name)']]],
-  ['setupcoroutineorderedbyname_62',['setupCoroutineOrderedByName',['../classace__routine_1_1Coroutine.html#a57603a24c66c747a44dafebc4af8ecf3',1,'ace_routine::Coroutine::setupCoroutineOrderedByName(const char *name)'],['../classace__routine_1_1Coroutine.html#a93c4fc4b3ddbffbac26c529860b9ec78',1,'ace_routine::Coroutine::setupCoroutineOrderedByName(const __FlashStringHelper *name)']]],
-  ['setvalue_63',['setValue',['../classace__routine_1_1Channel.html#ab950ccdf77fdfaeea67ac19031733229',1,'ace_routine::Channel']]],
-  ['setyielding_64',['setYielding',['../classace__routine_1_1Coroutine.html#a87f187a8d19348f7596033b1c86c85dd',1,'ace_routine::Coroutine']]],
-  ['status_65',['Status',['../classace__routine_1_1Coroutine.html#a2c0991faf233080e16bd81b94db0adbb',1,'ace_routine::Coroutine']]],
-  ['statusprintto_66',['statusPrintTo',['../classace__routine_1_1Coroutine.html#af5a9a15d88883683455625537e99fc19',1,'ace_routine::Coroutine']]],
-  ['suspend_67',['suspend',['../classace__routine_1_1Coroutine.html#ad919229c4f2f2c85d8d65238d19316f8',1,'ace_routine::Coroutine']]]
+  ['read_49',['read',['../classace__routine_1_1Channel.html#a931ce2344667a4b36e76bb4f8f3b70a0',1,'ace_routine::Channel']]],
+  ['reset_50',['reset',['../classace__routine_1_1CoroutineTemplate.html#aac6c42a690d4f64942c7603b9890de9b',1,'ace_routine::CoroutineTemplate']]],
+  ['resume_51',['resume',['../classace__routine_1_1CoroutineTemplate.html#aab8a1eba56fb3d10273ce3c62263718b',1,'ace_routine::CoroutineTemplate']]],
+  ['runcoroutine_52',['runCoroutine',['../classace__routine_1_1CoroutineTemplate.html#a3da6aa3220b54203f050dc97d51b7070',1,'ace_routine::CoroutineTemplate']]]
 ];