+++ docs/html/search/classes_0.js
@@ -1,6 +1,7 @@
 var searchData=
 [
-  ['channel_70',['Channel',['../classace__routine_1_1Channel.html',1,'ace_routine']]],
-  ['coroutine_71',['Coroutine',['../classace__routine_1_1Coroutine.html',1,'ace_routine']]],
-  ['coroutinescheduler_72',['CoroutineScheduler',['../classace__routine_1_1CoroutineScheduler.html',1,'ace_routine']]]
+  ['channel_68',['Channel',['../classace__routine_1_1Channel.html',1,'ace_routine']]],
+  ['clockinterface_69',['ClockInterface',['../classace__routine_1_1ClockInterface.html',1,'ace_routine']]],
+  ['coroutineschedulertemplate_70',['CoroutineSchedulerTemplate',['../classace__routine_1_1CoroutineSchedulerTemplate.html',1,'ace_routine']]],
+  ['coroutinetemplate_71',['CoroutineTemplate',['../classace__routine_1_1CoroutineTemplate.html',1,'ace_routine']]]
 ];