+++ docs/html/search/defines_1.js
@@ -1,6 +1,13 @@
 var searchData=
 [
-  ['extern_5fcoroutine_136',['EXTERN_COROUTINE',['../Coroutine_8h.html#a79f1cf897012bbc4df5565f61adc3028',1,'Coroutine.h']]],
-  ['extern_5fcoroutine1_137',['EXTERN_COROUTINE1',['../Coroutine_8h.html#ab92199c3991124cbb6e5e48e27827dae',1,'Coroutine.h']]],
-  ['extern_5fcoroutine2_138',['EXTERN_COROUTINE2',['../Coroutine_8h.html#af52549e71d4a6f7bd18c4688042776b6',1,'Coroutine.h']]]
+  ['coroutine_122',['COROUTINE',['../Coroutine_8h.html#ab22e58ab24b1908a7280f13603248ce3',1,'Coroutine.h']]],
+  ['coroutine1_123',['COROUTINE1',['../Coroutine_8h.html#a8d8219c8fc07650dfcdffbf83d800574',1,'Coroutine.h']]],
+  ['coroutine2_124',['COROUTINE2',['../Coroutine_8h.html#a04a0fdcb7293e51ddda298e237264c51',1,'Coroutine.h']]],
+  ['coroutine_5fawait_125',['COROUTINE_AWAIT',['../Coroutine_8h.html#a9f6714d9752740b6173b6afc4d108fa8',1,'Coroutine.h']]],
+  ['coroutine_5fbegin_126',['COROUTINE_BEGIN',['../Coroutine_8h.html#a7b7dc05639a37cb3f51bf348c68db283',1,'Coroutine.h']]],
+  ['coroutine_5fdelay_127',['COROUTINE_DELAY',['../Coroutine_8h.html#a1391503960a4b0c48f8200f1ffa0e305',1,'Coroutine.h']]],
+  ['coroutine_5fend_128',['COROUTINE_END',['../Coroutine_8h.html#a7e326a8df652c46c42a8d713ddc67bd7',1,'Coroutine.h']]],
+  ['coroutine_5floop_129',['COROUTINE_LOOP',['../Coroutine_8h.html#adad73df0853c1d3040b88b3cd0a04c0e',1,'Coroutine.h']]],
+  ['coroutine_5fyield_130',['COROUTINE_YIELD',['../Coroutine_8h.html#a682abe1669f9fe28af63d11da5421cbf',1,'Coroutine.h']]],
+  ['coroutine_5fyield_5finternal_131',['COROUTINE_YIELD_INTERNAL',['../Coroutine_8h.html#a53efafb6fd212ea352f5ad52ea9bc393',1,'Coroutine.h']]]
 ];