+++ docs/html/search/functions_8.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['_7ecoroutinetemplate_107',['~CoroutineTemplate',['../classace__routine_1_1CoroutineTemplate.html#a47cd05a5a9c1f63fb8b44d756aad4e7b',1,'ace_routine::CoroutineTemplate']]]
+];