+++ docs/html/search/searchdata.js
@@ -1,12 +1,12 @@
 var indexSectionsWithContent =
 {
-  0: "acefgiklrsw~",
+  0: "acefgiklmrsw~",
   1: "c",
   2: "c",
-  3: "cgilrsw~",
-  4: "k",
-  5: "s",
-  6: "cefg",
+  3: "cgilmrsw~",
+  4: "km",
+  5: "cs",
+  6: "acefg",
   7: "a"
 };