+++ examples/MemoryBenchmark/collect.sh
@@ -16,6 +16,7 @@
 set -eu
 
 PROGRAM_NAME='MemoryBenchmark.ino'
+NUM_FEATURES=8 # excluding FEATURE_BASELINE
 
 # Assume that https://github.com/bxparks/AUniter is installed as a
 # sibling project to AceRoutine.
@@ -49,7 +50,7 @@ function collect_for_board() {
     local board=$2
     local result_file=$3
 
-    for feature in {0..4}; do
+    for feature in $(seq 0 $NUM_FEATURES); do
         echo "Collecting flash and ram usage for FEATURE $feature"
         sed -i -e "s/#define FEATURE [0-9]*/#define FEATURE $feature/" \
             $PROGRAM_NAME