+++ examples/MemoryBenchmark/generate_readme.py
@@ -6,6 +6,8 @@
 
 from subprocess import check_output
 
+attiny_results = check_output(
+    "./generate_table.awk < attiny.txt", shell=True, text=True)
 nano_results = check_output(
     "./generate_table.awk < nano.txt", shell=True, text=True)
 micro_results = check_output(
@@ -38,7 +40,7 @@ calculated flash size can jump around in unexpected ways.
 
 **NOTE**: This file was auto-generated using `make README.md`. DO NOT EDIT.
 
-**Version**: AceRoutine v1.2
+**Version**: AceRoutine v1.3
 
 **Changes**:
 
@@ -51,6 +53,39 @@ calculated flash size can jump around in unexpected ways.
       linked list
     * `CoroutineScheduler::setupScheduler()` simplified due to immutable linked
       list
+* v1.3
+    * Remove virtual destructor on `Coroutine` class. Reduces flash memory
+      consumption by 500-600 bytes on AVR processors, 350 bytes on SAMD21, and
+      50-150 bytes on other 32-bit processors. The static memory is also reduced
+      by 14 bytes on AVR processors.
+    * Replace clock ticking virtual methods (`coroutineMicros()`,
+      `coroutineMillis()`, and `coroutineSeconds()`) with static functions that
+      delegate to `ClockInterface` which is a template parameter. Saves only
+      0-40 bytes of flash on on AVR processors, but 100-1500 bytes of flash on
+      32-bit processors.
+    * Add benchmark for 'One Delay Function' and 'Two Delay Functions` which use
+      functions with a non-blocking if-statement to implement the functionality
+      of a `COROUTINE()` that loops every 10 milliseconds.
+    * Remove `COROUTINE_DELAY_SECONDS()` functionality. Saves about 200 bytes on
+      AVR processors, mostly from the removal of `udiv1000()` which takes almost
+      180 bytes. Replacing with native `/1000` does not help much because native
+      long division consumes about 130 bytes and is 3X slower on AVR processors.
+    * Remove `COROUTINE_DELAY_MICRO()` functionality. Saves about 15-20 bytes
+      of flash memory per coroutine on AVR. Saves 80-100 bytes plus 20-30 bytes
+      of flash per coroutine on 32-bit processors.
+    * Add `Blink Function` and `Blink Coroutine`, 2 implementations of the same
+      asymmetric blink functionality where the HIGH level lasts for a different
+      duration than the LOW level.
+    * Remove `Coroutine::getName()`, `Coroutine::mName()`, and
+      `Coroutine::setupCoroutine()`. The human-readable name is no longer
+      retained. Saves 10-30 bytes of flash and 3 bytes of static memory per
+      coroutine instance for AVR; 10-40 bytes and 8 bytes of static memory per
+      instance on 32-bit processors.
+    * Use 2 different `Coroutine` subclasses, instead of 2 instances of the same
+      `Coroutine` subclass, for the "Scheduler, Two Coroutines" benchmark . This
+      makes it more functionally equal to the "Two Coroutines" benchmark which
+      automatically creates 2 different subclasses through the `COROUTINE()`
+      macro. Increases flash memory by 80-130 bytes across the board.
 
 ## How to Generate
 
@@ -89,19 +124,35 @@ will be invoked by the following command:
 $ make README.md
 ```
 
-## Functionality
+## Results
 
 * Baseline: A program that does (almost) nothing
-* Coroutine (bare): A single `COROUTINE()` macro that does nothing.
-* Coroutine (LOOP,DELAY): A `COROUTINE()` macro that uses `COROUTINE_LOOP()`
-  and `COROUTINE_DELAY()` which are expected to used in the common case.
-* CoroutineScheduler (bare): A single `Coroutine` instance with a
-* `CoroutineScheduler`.
-* CoroutineScheduler (LOOP,DELAY): A single `Coroutine` instance that
-  uses `COROUTINE_LOOP()` and `COROUTINE_DELAY()` managed by a
+* One Delay Function: A single non-blocking delay function that waits 10 millis.
+* Two Delay Functions: Two non-blocking delay functions.
+* One Coroutine: One instance of `Coroutine` that waits 10 millis, executed
+  directly through `runCoroutine()`.
+* Two Coroutines: Two instances of `Coroutine` that wait 10 millis, executed
+  directly through `runCoroutine()`.
+* Scheduler, One Coroutine: One instance of `Coroutine` executed through the
   `CoroutineScheduler`.
+* Scheduler, Two Coroutines: Two instances of `Coroutine` executed through the
+  `CoroutineScheduler`.
+* Blink Function: A function that blinks the LED asymmetrically, with HIGH
+  lasting a different duration than LOW.
+* Blink Coroutine: A `Coroutine` that blinks asymmetrically, exactly the same as
+  the `blink()` function.
 
-## Arduino Nano
+### ATtiny85
+
+* 8MHz ATtiny85
+* Arduino IDE 1.8.13
+* SpenceKonde/ATTinyCore 1.5.2
+
+```
+{attiny_results}
+```
+
+### Arduino Nano
 
 * 16MHz ATmega328P
 * Arduino IDE 1.8.13
@@ -111,7 +162,7 @@ $ make README.md
 {nano_results}
 ```
 
-## Sparkfun Pro Micro
+### SparkFun Pro Micro
 
 * 16 MHz ATmega32U4
 * Arduino IDE 1.8.13
@@ -121,17 +172,17 @@ $ make README.md
 {micro_results}
 ```
 
-## SAMD21 M0 Mini
+### SAMD21 M0 Mini
 
 * 48 MHz ARM Cortex-M0+
 * Arduino IDE 1.8.13
-* Sparkfun SAMD Core 1.8.1
+* SparkFun SAMD Core 1.8.1
 
 ```
 {samd_results}
 ```
 
-## STM32 Blue Pill
+### STM32 Blue Pill
 
 * STM32F103C8, 72 MHz ARM Cortex-M3
 * Arduino IDE 1.8.13
@@ -141,7 +192,7 @@ $ make README.md
 {stm32_results}
 ```
 
-## ESP8266
+### ESP8266
 
 * NodeMCU 1.0 clone, 80MHz ESP8266
 * Arduino IDE 1.8.13
@@ -151,17 +202,17 @@ $ make README.md
 {esp8266_results}
 ```
 
-## ESP32
+### ESP32
 
 * ESP32-01 Dev Board, 240 MHz Tensilica LX6
 * Arduino IDE 1.8.13
-* ESP32 Boards 1.0.4
+* ESP32 Boards 1.0.6
 
 ```
 {esp32_results}
 ```
 
-## Teensy 3.2
+### Teensy 3.2
 
 * 96 MHz ARM Cortex-M4
 * Arduino IDE 1.8.13